Move mock strings into math-input.
